// functions/src/cloud_functions/bets/models/games/nfl_game.ts
export interface NflGame {
  GameKey?: string;
  SeasonType?: number;
  Season?: number;
  Week?: number;
  Date?: Date;
  AwayTeam?: string;
  HomeTeam?: string;
  AwayScore?: null;
  HomeScore?: null;
  Channel?: null | string;
  PointSpread?: number;
  OverUnder?: number;
  Quarter?: null;
  TimeRemaining?: null;
  Possession?: null;
  Down?: null;
  Distance?: null;
  YardLine?: null;
  YardLineTerritory?: null;
  RedZone?: null;
  AwayScoreQuarter1?: null;
  AwayScoreQuarter2?: null;
  AwayScoreQuarter3?: null;
  AwayScoreQuarter4?: null;
  AwayScoreOvertime?: null;
  HomeScoreQuarter1?: null;
  HomeScoreQuarter2?: null;
  HomeScoreQuarter3?: null;
  HomeScoreQuarter4?: null;
  HomeScoreOvertime?: null;
  HasStarted?: boolean;
  IsInProgress?: boolean;
  IsOver?: boolean;
  Has1stQuarterStarted?: boolean;
  Has2ndQuarterStarted?: boolean;
  Has3rdQuarterStarted?: boolean;
  Has4thQuarterStarted?: boolean;
  IsOvertime?: boolean;
  DownAndDistance?: null;
  QuarterDescription?: string;
  StadiumID?: number;
  LastUpdated?: Date;
  GeoLat?: null;
  GeoLong?: null;
  ForecastTempLow?: number;
  ForecastTempHigh?: number;
  ForecastDescription?: string;
  ForecastWindChill?: number;
  ForecastWindSpeed?: number;
  AwayTeamMoneyLine?: number;
  HomeTeamMoneyLine?: number;
  Canceled?: boolean;
  Closed?: boolean;
  LastPlay?: null;
  Day?: Date;
  DateTime?: Date;
  AwayTeamID?: number;
  HomeTeamID?: number;
  GlobalGameID?: number;
  GlobalAwayTeamID?: number;
  GlobalHomeTeamID?: number;
  PointSpreadAwayTeamMoneyLine?: number;
  PointSpreadHomeTeamMoneyLine?: number;
  ScoreID?: number;
  Status?: string;
  GameEndDateTime?: null;
  HomeRotationNumber?: number;
  AwayRotationNumber?: number;
  NeutralVenue?: boolean;
  RefereeID?: number;
  OverPayout?: number;
  UnderPayout?: number;
  StadiumDetails?: StadiumDetails;
}

export interface StadiumDetails {
  StadiumID?: number;
  Name?: string;
  City?: string;
  State?: string;
  Country?: string;
  Capacity?: number;
  PlayingSurface?: string;
  GeoLat?: number;
  GeoLong?: number;
  Type?: string;
}